Add options to mask seeds using the pin
This allows much greater security, since the seeds are no longer
trivially recoverable if this option is set without the pin.  Also
provides an even stronger security option to not validate the pin by not
storing the hash at all.  Instead, incorrect pins just produce incorrect
seeds and thus incorrect tokens, making it very difficult to recover the
original seeds.  (Which of course has its downsides too, if you forget
or mis-enter a PIN.  We could require the PIN to be entered twice in
this case instead.)